Add files via upload
In this  project, created a photo frame. The project will use HTML, CSS, and JavaScript to create a beautiful photo gallery in your browser. The page consists of a box that changes the containing text and two buttons. On click, the buttons show and hide a paragraph. Lastly, it contains four images displaying logos. The different page elements have numerous styles and colors, creating an attractive webpage.

The first things create an external JavaScript file and add it to the HTML file with the help of a <script> tag. Next,  added JavaScript events to the HTML elements that are linked to the JavaScript functions. The functions trigger whenever these events occur and perform specified actions.

The functions upDate(element) and unDo() alter the text shown in the blue box. The text in the box is either a predefined line or the alt text of the image the user hovers over. The functions openFunction() and closeFunction() show and hide the text box respectively. 
i used the document’s built-in method getElementById() to access the elements we wanted to get.